+++ app/src/main/java/eu/kanade/tachiyomi/data/track/myanimelist/MyAnimeListApi.kt
@@ -189,7 +189,7 @@ class MyAnimeListApi(private val client: OkHttpClient, interceptor: MyAnimeListI
             val isRereading = listStatus.isRereading
             status = if (isRereading) MyAnimeList.REREADING else getStatus(listStatus.status)
             last_chapter_read = listStatus.numChaptersRead.toFloat()
-            score = listStatus.score.toFloat()
+            score = listStatus.score.toDouble()
             listStatus.startDate?.let { started_reading_date = parseDate(it) }
             listStatus.finishDate?.let { finished_reading_date = parseDate(it) }
         }
